Threads that crash should not becoming joinable
If we get a trap or exception form inside of a thread main function we
should not mark the thread as joinable but instead just re-throw so that
the exception is transmitted back to the main thread and the whole
application is brought down.

Sadly this doesn't currently work if the main thread is block in
`pthread_join` (because the posted exception will never get recieved) so
this test currently runs with `PROZY_TO_PTHREAD`.

Tested both under node and in the browser since they have slightly
different worker models.